Almost done implementing wall and floor measurements fully.
Just got a small section to finish. I used the cobble walls that were already at the forefront of the upper section but when i got to the lower part as no forward walls had been put in i just placed wall measuremetns. They can just be replaced if needed